requirements.txt
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
idna==2.10
mysql-connector-python==8.0.22
protobuf==3.13.0
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.26.1
